extract dropdown list to its own file
